improve testing and examples
